Split protocol DTOs for AG14 fix
